+++ src/core_modules/capture-core-utils/validators/form/percentage.validator.js
@@ -3,7 +3,7 @@
  *
  * @export
  * @param {string} value
- * @returns
+ * @returns {boolean}
  */
 export const isValidPercentage = (value: string) =>
     /^(100|\d\d|\d)([,.]0*)?( *)?%?$/.test(value);