+++ src/core_modules/capture-core-utils/validators/form/phone.validator.js
@@ -4,6 +4,6 @@
  *
  * @export
  * @param {string} value
- * @returns
+ * @returns {boolean}
  */
 export const isValidPhoneNumber = (value: string) => /^[^a-zA-Z]+$/.test(value);