+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.component.js
@@ -3,9 +3,10 @@ import React from 'react';
 import type { ComponentType } from 'react';
 import i18n from '@dhis2/d2-i18n';
 import { spacersNum } from '@dhis2/ui';
+import { dataEntryIds, dataEntryKeys } from 'capture-core/constants';
 import withStyles from '@material-ui/core/styles/withStyles';
 import type { PlainProps } from './EnrollmentEditEventPage.types';
-import { pageMode, pageStatuses } from './EnrollmentEditEventPage.constants';
+import { pageStatuses } from './EnrollmentEditEventPage.constants';
 import { WidgetEventEdit } from '../../WidgetEventEdit/';
 import { WidgetError } from '../../WidgetErrorAndWarning/WidgetError';
 import { WidgetWarning } from '../../WidgetErrorAndWarning/WidgetWarning';
@@ -65,6 +66,7 @@ const EnrollmentEditEventPagePain = ({
     eventStatus,
     pageStatus,
     onEnrollmentError,
+    onCancelEditEvent,
 }: PlainProps) => (
     <OrgUnitFetcher orgUnitId={orgUnitId}>
         <TopBar
@@ -82,7 +84,7 @@ const EnrollmentEditEventPagePain = ({
         />
         <div className={classes.page}>
             <div className={classes.title}>
-                {mode === pageMode.VIEW
+                {mode === dataEntryKeys.VIEW
                     ? i18n.t('Enrollment{{escape}} View Event', { escape: ':' })
                     : i18n.t('Enrollment{{escape}} Edit Event', { escape: ':' })}
             </div>
@@ -96,6 +98,7 @@ const EnrollmentEditEventPagePain = ({
                             orgUnitId={orgUnitId}
                             enrollmentId={enrollmentId}
                             eventStatus={eventStatus}
+                            onCancelEditEvent={onCancelEditEvent}
                         />
                     )}
                     {pageStatus === pageStatuses.MISSING_DATA && (
@@ -108,7 +111,7 @@ const EnrollmentEditEventPagePain = ({
                     )}
                 </div>
                 <div className={classes.rightColumn}>
-                    <WidgetEventComment itemId={mode} dataEntryId="singleEvent" />
+                    <WidgetEventComment dataEntryKey={mode} dataEntryId={dataEntryIds.ENROLLMENT_EVENT} />
                     <WidgetError error={widgetEffects.errors} />
                     <WidgetWarning warning={widgetEffects.warnings} />
                     {!hideWidgets.feedback && (