+++ src/core_modules/capture-core/components/WidgetEventComment/WidgetEventComment.types.js
@@ -1,7 +1,7 @@
 // @flow
 
 export type Props = {|
-    itemId: string,
+    dataEntryKey: string,
     dataEntryId: string,
     ...CssClasses
 |};