+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/constants/additionalFilters.js
@@ -1,12 +1,12 @@
 // @flow
 export const ADDITIONAL_FILTERS = {
     programStage: 'programStage',
-    reportDate: 'reportDate',
-    eventStatus: 'eventStatus',
+    occurredAt: 'occurredAt',
+    status: 'status',
 };
 
 export const ADDITIONAL_FILTERS_LABELS = {
     programStage: 'Program stage',
-    reportDate: 'Report date',
-    eventStatus: 'Event status',
+    occurredAt: 'Report date',
+    status: 'Event status',
 };