+++ src/core_modules/capture-core/utils/routing/index.js
@@ -2,3 +2,4 @@
 export { useLocationQuery } from './useLocationQuery';
 export { getLocationQuery } from './getLocationQuery';
 export { buildUrlQueryString } from './buildUrlQueryString';
+export { shouldUseNewDashboard } from './newDashboard';